Fixes 2 - Add management command check_contest to anycontest app
